Use elm.json files to understand app dependencies
Now that Elm 0.19 stores dependencies outside of
the elm-stuff directory and instead stores them
in a per-user cache (e.g. ~/.elm), it is necessary
to parse the `elm.json` file in order to understand
which packages an application depends on in order
to distinguish between other packages that should
not be in scope.

This implementation is merely meant as a starting
point. Current limitations:

- cannot attach multiple `elm.json` files
- cannot attach a package's `elm.json` files
  (only 'application' projects are supported right now)
- there is no UI for seeing which projects are attached
- the PsiReference resolve system does not fully leverage
  the information from the `elm.json` file. Specifically,
  it does not treat 'test-directories' differently, nor
  does it use 'source-directories' to create src roots,
  nor does it use a package's 'exposed-modules' to
  determine which modules are visible from outside of a
  package.

I added support for configuring the location of the
Elm binary, but right now it is not being used.
In the very near future we will need this in order to
properly construct a versioned path into the per-user
cache. And it will be useful for adding new features
like adding a run-config to launch Elm Reactor.