Allow attaching "package" project `elm.json` file (#69)
This registers the package dependencies as external library roots. 
Unlike "application" projects, we cannot easily obtain:

- the exact version numbers used
- the transitive dependencies

So this PR just registers the package's direct dependencies and will
use an arbitrary version that it finds in ~/.elm that satisfies its own 
version constraint (but the full dependency graph).

More work needs to be done to make this better. Ideally the Elm 
toolchain itself would be able to describe how it would resolve a 
package's dependency graph.