Fix import intention for type annotation parameters
The problem was that in this case, the cursor is on a whitespace element adjacent to the `ElmUpperPathTypeRef` rather than the name itself, so the parent is the type annotation.

We should never try to import he function you're annotating, so I disallow it. This also fixes the case where the cursor is over the function name in the annotation.

Fixes #56