Improve lexing of partial string literals
The lexer currently requires the closing quote for string literals. So while typing, the string is not highlighted. In some cases, the string will not be highlighted correctly even after typing the closing quote. This happens when the lexer emits tokens after the string error token. In that case, the partial lex doesn't look at the start of the string again, so more error tokens are emitted.

This PR changes the lexer to make closing quotes in string literals optional. An annotator was added to report the missing closing quotes, although since all strings literals in elm are multi-line, the error is always reported at the end of the file.

Fixes #78