Allow FieldAccess as target for function calls
We were incorrectly rejecting syntax like the following:

```elm
input =  { x = \() -> () }
valueRef = { foo = input.x (), bar = () }
                   -------
```

The issue was that `FieldAccess` wasn't an option in the `FunctionCallTarget` rule.

In order to get this to work, I cleaned up the hack that required `FieldAccess` be a higher priority than `FunctionCall`. Now we do a small lookahead in all the whitespace-sensitive rules to make sure they don't match on part of a dotted id.