Fix tests for elements which have multiple references
The problem was that for something like `Foo.bar`, there are two
references: one on the `Foo` part and the other on the `bar` part.
Our old testing system would always find the latter. This was resulting
in passing tests for the Cmd and Sub alias default import when they were
actually broken.