Fix references to Cmd and Sub
This fixes a regression in #92 that caused implicit references to `Cmd` and `Sub` not to resolve. These two values are special-cased in `moduleLookupHack` when creating the module definition, but not the dependency exposed modules, so `ElmProject.exposes` was returning false because `exposedModules` contained `Platform.Cmd`, but the module declaration name was just `Cmd`

I added the hack to the `exposedModules`, but I don't know if that's the correct fix. Maybe the hack is no longer necessary?