Smarter insertion of virtual end decl tokens (#109)
Previously we were inserting them immediately after the last
code token, but if there was trailing whitespace or trailing
comment on that expression, that whitespace/comment would end
up AFTER the virtual end decl. Which is not desirable because
it logically belongs to the decl.

The Psi tree after this change is still not great in these
cases: the trailing whitespace is a _sibling_ of the value declaration.
But at least now it comes _before_ the virtual end decl.