Allow filtering annotations and module declarations from Find Usages (#118)
This PR extends the import filtering rule to also exclude usages in `exposing` clauses and type annotations. The latter is semantically a bit kludgy, but this allows a user to filter the Find Usages panel for to only places where a function is actually called. As before, the entire rule can be toggled on and off.