Work around unresolved type aliases when infering type epxressions (#132)
The Elm core code has a lot of type aliases the reference synthetic classes:

```
type alias Html msg = VirtualDom.Node msg
```

The `VirtualDom` package is JS code, so the reference doesn't resolve. We currently treat the type alias as unknown in these cases. This PR chages to behavior to treat the alias as a union declaration when the reference in unresolved.