Make parenthesized expressions a real Psi element
This is necessary to enable GrammarKit's expression parser mode (Pratt)
because every child expression rule must be public.

But it's also useful in general as we can't just treat parens as sugar.
For instance, if we were to add functionality in the future that removes
unnecessary parens, you would want to work with the parens.