Importing a custom type variant constructor should import all variants
Starting in Elm 0.19, it is no longer valid to import individual variants
of a custom type. This change makes it so that when we add an import
for a single variant, it also exposes all of the other variants.

For Elm 0.18 users, this may be a minor inconvenience, but it's easy
enough to workaround.