Support more expression types in Smart Enter Processor (#205)
Add "smart-enter" support for

- Individual case branches
- if/then expressions
- let/in expressions
- value declarations (top level and inside let/in expressions)

The value declaration support is a bit hacky due to the way our grammar is structured. The `ValueDeclaration` rule pins on `=`, so a top-level identifier just parses as an error. We can't just pin on the id, since that breaks type annotations. We can change move `TypeAnnotation` to be higher priority than `ValueDeclaration`, but that still causes some partial programs to parse weirdly. Since there's no `VIRTUAL_OPEN_DECL` token, we don't know whether an id is on a line by itself, so sometimes a `ValueDeclaration` would start on the same line as other code.

There's probably a solution that involves moving the annotation and declaration into the same rule, but I'm not going to work on that for now.