Cache the Elm version used by the 'check toolchain' notification (#227)
It looks like there are some cases where IntelliJ aggressively invalidates the notification panel associated with each file. And every time this happened, we were synchronously invoking the external Elm process to check its version.

Now we cache the Elm compiler version and invalidate it whenever there is a change to the workspace or its settings (which includes the path to the Elm compiler).