Make function parameter rename a lot faster
Previously, when renaming a function parameter with a common name like
`index`, a "Finding Usages" dialog and progress bar would appear since
it was scanning across 10's or 100's of files. But now it only needs
to scan the text within the function body.

I found this trick in intellij-rust.