Fix a "use scope" bug for destructured patterns (#239)
Commit db2e23c changed the way that Find Usages works for `ElmLowerPattern`.
I thought this was constrained to just function parameters, but it also
affected things like `let` destructuring. For example:

f p =
     let
          (x, y) = p
     in
     x

Here, the usage scope for `x` includes the entire `let` expression.

The fix was to relax the scope a little bit. And I added some more tests so that we can catch this in the future.