Improve record completion in functions with syntax errors. (#245)
In the large majority of cases when trying to complete record fields, the target is a simple value reference, usually to a function parameter. If we resolve that reference, it often resolves to an element for which we can infer the type. In the case of chained field access, we can resolve the base type and walk the `TyRecord` to get the final field ty.

We now bind function parameters as long as there are no syntax errors in the parameters themselves, even if there are errors in the function body.