Clear the PsiReference cache whenever the Elm workspace changes (#275)
* Clear the PsiReference cache whenever the Elm workspace changes

The PsiReference cache is automatically cleared by IntelliJ whenever
a change occurs to any Psi in your project (e.g. typing some code).
And that's good. But we also need to be able to clear the cache
whenever an external change to the ElmProject occurs.

For instance, a user might have some imports with unresolved references
to packages which have not yet been installed. They switch over to their
terminal and use Elm's package installer to install the packages. Elm
updates the `elm.json` file, and our plugin notices the change and loads
the new dependency. Now the import which was unresolved should be able
to be resolved. But the old ref result was cached, so we need to clear
the cache.

* Invalidate the caches that depend on modification trackers

e.g. CachedValuesManager used by ModuleScope