Freshen all vars retrieved from a cache during inference.
Fixes #269

`Nothing` is of type `Maybe a`. In that issue, `a` was being unified to `String` correctly, but since both `a` variables were the same (because we were only freshening function return values), the second `Nothing` was already fixed to `String`, rather than being a unique variable that could be unified with `Int`.