Show nice errors from Elm compiler (#332)
There is now a tool window which shows the error output from the Elm compiler. The user can either press the build button or bind it to a keyboard shortcut to recompile their Elm app. The errors are shown with colored highlighting and clicking them navigates to the location in the source code.