Reduce memory usage of type inference (#422)
Previously, TypeReplacement would always do a deep copy of an Ty being replaced. This caused every cached Ty, including the Tys for each subexpression, to be unique objects on the heap. For large files with deeply nested types, this could consume a significant amount of memory.

With this PR, TypeReplacement returns the original Ty if there was no replacement necessary. This optimization is granular, so even if some parts of a Ty need to be replaced, other parts may be reused.

I also noticed that there were a large amount of empty ArrayLists on the heap. Many of Kotlin's collection functions optimize empty lists to a static singleton, but `map` some of the other transform functions don't. I assume this is an oversight, so I added the optimization to the TypeReplacement code.