Fix Pico injection exception on ElmTestRunConfigurationProducer
Raise minimum IntelliJ version to 2019.1 because we have to subclass
LazyRunConfigurationProducer, which was introduced in 2019.1. Otherwise,
we would have to create multiple builds to support 2018, and I just
don't have the time to do that. Sorry.

I also have decided to introduce an upper-bound on the IntelliJ
compatibility version since their API is breaking more than I expected.