Fix case where record field reference wouldn't resolve
To my surprise, it turns out that if you try _really_ hard, you can in fact assign an extension record to a concrete record before the concrete record is unified with a type annotation. This commit now tracks replacements for that case.

Note that this didn't affect inference correctness, it just prevented us from propagating field references.