Resolve record pattern references
This took two changes:

- Make ElmLowerPattern a reference element. It was already a named element; now it's both.
- Make LexicalValueReferences pass through parameters that are also references

The second change is necessary because otherwise find usages / renaming a record field would not find ElmValueExpr elements, since they would only resolve to the parameter rather than the field. Unfortunately, that change means that parameters that resolve to another element will always be marked as unused. I worked around that by ignoring those elements in the unused symbol inspection. It also means that we can't generate documentation for record pattern parameters that reference records directly inside type annotations (it still works for references to aliased records).

Given the imperfect nature of this fix, you may want to reject this PR. It seems that this feature is more useful than the ones it breaks, but that's up to you.