Use type annotations when inferring recursive functions.
Currently, we are returning TyUnknown for all referenced recursive value declarations to avoid infinite loops. This is overly cautious and results in some false negatives.

This commit uses type annotations for recursive functions when available, since there's no risk of a loop in that case. We still return TyUnknown when no annotation is present.