Fix replacement when freezing records with no other changes (#507)
A record's field reference table wasn't participating in equality. This is correct with respect to the fields themselves, but since the frozen status was not part of equality, the replacement code would return the unfrozen record when no other replacement were made. This caused unfrozen tables to be cached, leading to ConcurrentModificationExceptions.