Add real-world tests (#513)
* Add a test that runs inspections against real-world Elm projects
* Add highlighting performance test

All of these tests are optional. They will bail out if it cannot find the real-world Elm project on disk.

The performance test just measures current performance. It does not actually verify anything. There is no baseline and it does not lock the commit of the real-world Elm project that it's testing. This is meant to be a starting point for further automation.