Fix references to union constructors in arguments to other constructors in patterns
The grammar was previously parsing these as `ElmUpperCaseQID`, which is not a reference element. This PR adds a new element types `ElmNullaryConstructorArgumentPattern`, which wraps the QID with a reference element.

I made the name long rather something like `ElmUpperPattern`, because it's a very specific element. A pattern like `Nothing ->` will still parse as `ElmUnionPattern`. This is only used in patterns like `Just Nothing ->`, where the whole pattern will be an `ElmUnionPattern`, and the `Nothing` will now be `ElmNullaryConstructorArgumentPattern`.