Fix stack overflow in type inference of some malformed programs
In certain invalid programs, it was possible to assign a flex var to a rigid var in a way to would cause us to incorrectly store the replacement from rigid to flex rather than the other way around.

This commit contains two tests. Before this fix, the first test would incorrectly report a self-referential type, and the second would throw a StackOverflowException.