Add color swatch line markers (#569)
This PR adds a marker to lines containing expressions with colors. Clicking the marker will open a color picker, and chossing a color from that picker will update the expression, keeping whatever format and units currently exists.

I added support for colors in two formats:

## String literals

And string that contains a CSS hex, hsl(a), or rgb(a) color will work. The string can contain other characters. I chose not to limit this to arguments to know functions, since these formats are fairly unique, so it seems that the risk of false positives are low. On the other hand, because I didn't limit which strings can contain a color, I chose not to support CSS color keywords such as `"red"` or `"rebeccapurple"`, since I think it would lead to noticeable false-positives. Like all decisions, this is open for discussion.

## Function calls

I support a specific set of functions when called with number literals. I do not attempt to resolve references if the arguments to these functions are in a variable or other expression. I support most of the color functions from [elm-css](https://package.elm-lang.org/packages/rtfeldman/elm-css/latest/Css#rgb) and [elm-ui](https://package.elm-lang.org/packages/mdgriffith/elm-ui/latest/Element#rgb).

The following functions are supported:

- `rgb`, with either `Int` arguments (like in elm-css) or `Float` arguments (like elm-ui)
- `rgba`
- `rgb255`
- `rgba255`
- `hsl`
- `hsla`

The marker provider only looks at the function name and arguments. This mean we don't need to resolve any references in the marker provider, and allows us to support identical functions in other libraries (there are a number of Elm packages that copy these functions).

## Limitations

### Multiple colors in a string

Even though IntelliJ supports multiple color markers per line, due to the way `ElementColorProvider` is designed, each PsiElement can only have one color. This prevents us from showing more than one marker for each string literal, even if the colors are on separate lines.

A possible workaround would be to add a color to the `OPEN_QUOTE` and `CLOSE_QUOTE` elements in addition to the existing `REGULAR_STRING_PART` element. This would allow us to have at least three colors per string literal.

### Rounding

When a user chooses a color from the picker, floating point number might not match the chosen color exactly. `ElementColorProvider` gives us the user's choice as a `java.awt.Color`, which only has 8-bit color channels (including alpha) rather than floating point numbers.

For example, if the user chooses an alpha of `0.5`, `java.awt.Color` rounds this to 127, which is `0.498 * 255`, so we have no way of knowing that the user actually selected 0.5 instead. I'm open to suggestions on how to improve this situation.