Add options to disable individual line marker types
This consolidates all line marker providers into a single LineMarkerProviderDescriptor implementation that calls out to the individual providers and handles registering and checking options, as well as the boilerplate for `collectSlowLineMarkers`. Individual providers can just implement `getLineMarkerInfo`. Providers are created for each call to `collectSlowLineMarkers`, so they can have internal state if they need to.

Fixes #590