Optimize reference resolve (#585)
## ModuleScope

Although these loops were already cached, they were cached as a list of elements. Most reference implementations that used this scope would copy the returned list to an array (in `getVariants`), then filter that array down the the element that matches their name (in `resolveInner`).

The elements are now cached as both an array and a map of name to element, so now we don't make any copies, and `resolveInner` can do an O(1) map lookup. As an additional benefit, we only look up the element's `referenceName` once, where we were previously looking it up for each element in the array being searched.

## ImportScope

These results are now cached in a form similar to the form used in `ModuleScope`.

More importantly, calculating the values has been improved. Previously, calculating exposed names involved resolving the reference of each name in the exposing list. To do so, we would walk up from the reference to the file, then call a `ModuleScope.getDeclared*` function and filter the list for a matching name. But since all the exposed elements are defined in the same file, we now only call `ModuleScope` once and filter all the declarations in one pass. 

## ExpressionScope

This scope is very hot, but isn't amenable to caching due to the fact that the tree walk is unique for each reference, and retrieving a cached value is slow enough that caching intermediate results (e.g. `ElmValueDeclaration.declaredNames`) is a pessimization.

The main improvement to this scope is making the tree walk lazy so that e.g. resolving a reference to a function parameter doesn't need to call `ModuleScope.getVisibleValues`.

## GlobalScope

Since these results are the same for all files in a project, the results are now cached on the `ElmProject`. Additionally, the loops benefit from the improvements to `ModuleScope`

## Microoptimizations

I'm not going to list every change here, but there are a number of places that would walk over the same part of the Psi tree repeatedly. For example, `ElmUpperCaseQID.fillName`, which is a hot function, would gather all it's children into a list three times (once each for `isQualified`, `qualifierPrefix`, and `refName`). Now that function just calls `text` if it isn't a stub.