Fix a bug involving destructured assignment
This bug was causing false positives in the 'conflicting name' inspection.

e.g.
bar viewer =
    let
        (x, y) =
            case viewer of
                Editing str ->
                    (str, a)

                Sending str ->
                    (str, b)
    in
    x ++ y

The problem was that it was descending into the right-hand-side
of the destructured assignment and lifting out the `str` identifiers
in the case branches. And so it thought that `str` was declared twice
in the same scope.