Fix an exception when elm-format runs on-save when Editor not found
I didn't realize that a Document does not always have an Editor when
the on-save hook runs. Consider the following:

1. modify the contents of Foo.elm
2. close the editor tab for that file
3. change OS window manager focus from IntelliJ to another app
4. observe that the format-on-save hook runs but no Editor is available

The whole point of finding the editor was just to make it easier to show
all syntax errors using the `ElmBuildAction`. Since this is optional,
I decided to just get rid of the code path rather than try to find
a more robust solution.