Restricts the InlineDebugIntention from showing in several cases
To provide a better user experience, the intention should not be shown as an option to the user, if it cannot do something valuable.
This commit restricts the intention from showing in these cases:

 * On string, numeric, and char literals (unless the literal is part of a binary operator expression, in which case, THAT expression is logged).
 * On calls and arguments to Debug functions

This is also a prelude of sorts to having an intention for removing Debug.log statements.
Detecting if the given statement is part of a Debug.log call will require some of the same logic.